added confirmation for delete

+++ client/src/pages/AdminUsersList.js
@@ -19,7 +19,9 @@ const AdminUsersList = ({ history }) => {
 	const { userInfo } = userLogin;
 
 	const deleteHandler = (id) => {
-		dispatch(deleteUser(id));
+		if (window.confirm('Are you sure ? ')) {
+			dispatch(deleteUser(id));
+		}
 	};
 
 	useEffect(() => {
